requirements.txt
pymysql
pyyaml
tqdm